# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API v1 Server"
        }
    ],
    "paths": {
...
        "/api/v1/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/prouct": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store new product",
                "description": "Returns product data",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/product/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product information",
                "description": "Returns product data",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update existing product",
                "description": "Returns updated product data",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete existing product",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/product?product={product}&category={category}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product(s) information",
                "description": "Returns product(s) data",
                "operationId": "getProductsByName",
                "parameters": [
                    {
                        "name": "product",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateProductRequest": {
                "title": "CreateProductRequest",
                "description": "Create Product request body data",
                "required": [
                    "category_id",
                    "name",
                    "description",
                    "final_price"
                ],
                "properties": {
                    "category_id": {
                        "title": "Product Category ID",
                        "description": "ID of the category",
                        "example": "lorem"
                    },
                    "name": {
                        "title": "Product name",
                        "description": "Name of the product",
                        "example": "Lorem Ipsum Dolor"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the product",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.."
                    },
                    "old_price": {
                        "title": "Old price",
                        "description": "Old price of the product",
                        "example": "999.99"
                    },
                    "final_price": {
                        "title": "Final price",
                        "description": "Final price of the product",
                        "example": "1120"
                    }
                },
                "type": "object",
                "nullable": [
                    "old_price"
                ]
            },
            "UpdateProductRequest": {
                "title": "UpdateProductRequest",
                "description": "Update Product request body data",
                "required": [
                    "category_id",
                    "name",
                    "description",
                    "final_price"
                ],
                "properties": {
                    "category_id": {
                        "title": "Product Category ID",
                        "description": "ID of the category",
                        "example": "lorem"
                    },
                    "name": {
                        "title": "Product name",
                        "description": "Name of the product",
                        "example": "Lorem Ipsum Dolor"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the product",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.."
                    },
                    "old_price": {
                        "title": "Old price",
                        "description": "Old price of the product",
                        "example": "999.99"
                    },
                    "final_price": {
                        "title": "Final price",
                        "description": "Final price of the product",
                        "example": "1120"
                    }
                },
                "type": "object",
                "nullable": [
                    "old_price"
                ]
            },
            "ProductResource": {
                "title": "ProductResource",
                "description": "Product resource",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "old_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "final_price": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    },
...
}